Refactor App structure
